pos-print: pass department through receipt items
Different department now could be used for each of the receipt items.
Note that if no value is provided then a default one of "0" will be used
by Printer instance of FP705.

For example the structure becomes:
```
"receiptItems": [
  {
    "name": "Item1",
    "quantity": 2.0,
    "price": 1.0,
    "department": "0"
  }
]
```
where the department value is passed in the "registration of sale"
command.

Supported Printer: FP705